TEncSearch: Fix compile error due to char array subscript.
Sample compile errors in Linux64/GCC4.8.1:

$HEVC/source/Lib/TLibEncoder/TEncSearch.cpp:2134:75: error: array subscript has type ‘char’ [-Werror=char-subscripts]
     x265::pixelcmp sa8d = x265::primitives.sa8d[g_aucConvertToBit[uiWidth]];
                                                                           ^
$HEVC/source/Lib/TLibEncoder/TEncSearch.cpp: In member function ‘Void TEncSearch::estIntraPredQT(TComDataCU*, TComYuv*, TComYuv*, TShortYUV*, TComYuv*, UInt&, Bool)’:
$HEVC/source/Lib/TLibEncoder/TEncSearch.cpp:2206:79: error: array subscript has type ‘char’ [-Werror=char-subscripts]
         x265::pixelcmp sa8d = x265::primitives.sa8d[g_aucConvertToBit[uiWidth]];
                                                                               ^

+++ source/Lib/TLibEncoder/TEncSearch.cpp
@@ -2131,7 +2131,7 @@ Void TEncSearch::preestChromaPredMode(TComDataCU* pcCU,
     UInt  uiMaxMode   = 4;
     UInt  uiBestMode  = MAX_UINT;
     UInt  uiMinSAD    = MAX_UINT;
-    x265::pixelcmp sa8d = x265::primitives.sa8d[g_aucConvertToBit[uiWidth]];
+    x265::pixelcmp sa8d = x265::primitives.sa8d[(int)g_aucConvertToBit[uiWidth]];
     for (UInt uiMode  = uiMinMode; uiMode < uiMaxMode; uiMode++)
     {
         //--- get prediction ---
@@ -2203,7 +2203,7 @@ Void TEncSearch::estIntraPredQT(TComDataCU* pcCU,
         UInt uiStride      = pcPredYuv->getStride();
         UInt uiRdModeList[FAST_UDI_MAX_RDMODE_NUM];
         Int numModesForFullRD = g_aucIntraModeNumFast[uiWidthBit];
-        x265::pixelcmp sa8d = x265::primitives.sa8d[g_aucConvertToBit[uiWidth]];
+        x265::pixelcmp sa8d = x265::primitives.sa8d[(int)g_aucConvertToBit[uiWidth]];
 
         Bool doFastSearch = (numModesForFullRD != numModesAvailable);
         if (doFastSearch)
